+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Add an on.change meta node
 - Add a write2x option to save files also in double resolution
 - Add some tests
+- Add class Disposed to manage disposed nodes and test disposal and recreation of nodes cleanly
 
 ### Changed